LayoutTests/inspector/debugger/script-formatter-breakpoints-expected.txt

Tests the script formatting is working fine with breakpoints.

Debugger was enabled.

Running: testBreakpointsInOriginalAndFormattedSource
Script execution paused.

Dump breakpoint sidebar pane:

    script-formatter-breakpoints.html:11
    var f=0;
Script execution resumed.
Page reloaded.
Navigated to test-navigation.html
Page reloaded.
Navigated back to test page.

Dump breakpoint sidebar pane:

Script execution paused.

Dump breakpoint sidebar pane:

    script-formatter-breakpoints.html:25
    return 0;
Script execution resumed.

Dump breakpoint sidebar pane:

    script-formatter-breakpoints.html:11
    var f=0;
Debugger was disabled.
